+++ R/betaSandwich-diff-beta-ci-dot.R
@@ -1,29 +1,29 @@
 #' Confidence Intervals for
-#' Standardized Regression Coefficients
+#' Differences of Standardized Regression Slopes
 #'
 #' @author Ivan Jacob Agaloos Pesigan
 #'
 #' @return Returns a matrix of
-#'   standardized regression slopes,
+#'   differences of standardized regression slopes,
 #'   standard errors,
 #'   test statistics,
 #'   p-values,
 #'   and
 #'   confidence intervals.
 #'
-#' @param object Object of class `betasandwich`.
+#' @param object Object of class `difbetadelta`.
 #' @param alpha Numeric vector.
 #'   Significance level.
 #'
 #' @family Beta Sandwich Functions
-#' @keywords betaSandwich ci internal
+#' @keywords betaSandwich dif ci internal
 #' @noRd
-.BetaCI <- function(object,
-                    alpha = c(0.05, 0.01, 0.001)) {
+.DiffBetaCI <- function(object,
+                        alpha = c(0.05, 0.01, 0.001)) {
   stopifnot(
     methods::is(
       object,
-      "betasandwich"
+      "diffbetasandwich"
     )
   )
   return(
@@ -32,8 +32,7 @@
       se = sqrt(diag(object$vcov)),
       theta = 0,
       alpha = alpha,
-      z = FALSE,
-      df = object$lm_process$df
+      z = TRUE
     )
   )
 }