+++ NEWS.md
@@ -1,8 +1,12 @@
+# betaSandwich 1.0.4
+
+* Added the `DiffBetaSandwich()` and the `RSqBetaSandwich()` functions.
+
 # betaSandwich 1.0.3
 
 ## Patch
 
-* Added the `BetaADF` function.
+* Added the `BetaADF()` function.
 
 # betaSandwich 1.0.2
 
@@ -14,7 +18,7 @@
 
 ## Patch
 
-* Added `nas1982`.
+* Added the `nas1982` data set.
 * Minor edits to methods.
 
 # betaSandwich 1.0.0